requirements.txt
fastapi==0.123.5
numpy==1.26.4
pandas==2.3.3
Pillow==11.1.0
plotly==6.3.0
pydantic==2.12.5
Requests==2.32.5
scikit_learn==1.7.2
streamlit==1.45.1
tensorflow==2.19.0
uvicorn==0.38.0
python-multipart
